hwspi.settings must be a pointer to avoid undefined behavior.
AVR-g++ cannot perform the placement-new to initialize union
members, and it is a platform-defined class that is not trivial.

This member was never really safe because we don't know its
constructor-enforced invariants, so assigning to it is generally unsafe.
Indeed, particle-io defines SPISettings with virtual functions, private
members, and a base class.